Scripted install of Scalar and supporting components
Scalar requires several components to be configured in order to have a fully
functional installation. This includes components included in the base Scalar
installation (Scalar, Git), as well as some other components not included in
the base installation at this time (Watchman, GCM Core).

This change introduces a couple of scripts to 1) Gather the various support
components into a common location and 2) install Scalar and the supporting
components. The install script depends on HomeBrew being installed on the
machine to download and install Watchman. If HomeBrew is not installed, the
install script will fail with directions to install HomeBrew. We expect the
automation machines to already have HomeBrew installed. We can look at
alternative mechanisms for distributing Watchman in the future if necessary.

CreateScalarDistribution.sh: Gathers components required for a scripted install
    and place them in a common location (BuildOutput/Scalar.Distribution by
    default). This script will also generate a driver script to configure the
    environment and run ScalarInstallHelper.sh

ScalarInstallHelper.sh: Given a directory containing Scalar installer packages,
    install Scalar and supporting components. This will also use homebrew
    to download and install watchman.

These scripts can be used to set up a Scalar environment on an automation