ScalarInstall: more robust script handling
The pattern of the Scalar distribution writing out a "driver" script sets
environment variables and then calls another script can have issues when
calling the driver script from a directory other than the one that contains the
scripts. Instead, use a template for the installation script and set the
desired configuration by token replacement, which should be more robust.