Remove mapping.dat from .scalarCache
In the future there will not be a mount verb or process to maintain
the mapping.dat file that tracks the cache subdirectories of
.scalarCache.  This file had been used as a way to ensure that if
a user clones the same repo using different URLs that both clones
will share the same local cache.

Rather than relying on mapping.dat (and the cache server returned
by the /gvfs/config endpoint) to determine the local cache key
Scalar will now use the repo ID from the /vsts/info endpoint as the
local cache key or the SHA1 hash of the repo's URL if the info
endpoint is not supported or not available.  If a user has
multiple clones of the same repo they will continue to use the same
local cache key/directory.