Add maintenance verb and remove dehydrate verb
Add a maintenance verb that can be used to run
the LooseObjects, PackFileMainteance, and PostFetch
maintenance tasks.  For now these tasks are only called
by the functional tests, but eventually they will be
called by Scalar.Service as well.

The dehydrate verb was only around to facilated the
functional tests.  It is not removed in favor of the
maintenance verb.