maintenance: move maintenance task management to service
Previously, Scalar maintenance tasks were performed in the
Scalar.Mount process.  As a step toward completely removing
the mount process (and its associated verbs) the changes in
this commit move the scheduling and calling of maintenance
tasks into the service.

The service runs the maintenance tasks by way of the
'scalar maintenance' verb, and it does so for all of the
that the current user has registered.

With this change the 'mount' and 'unmount' verbs serve only
to register and unregister the repo with the service.
Deprecation of these verbs, as well as the mount process itself,
will be performed in a subsequent patch series.