Delete unused verbs: log and repair
LogVerb: this verb just isn't useful. We are going to change the log
behavior anyway.

RepairVerb: VFS for Git used to modify the index and HEAD in a way that
could lead to corruption. The repair verb was necessary to fix users
that hit those bugs. We should not need repairs in Scalar.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>