FetchCommitsAndTreesStep: wait longer if no cache server
We previously would not schedule the FetchCommitsAndTreesStep at all if
there was no cache server. With the maintenance tasks moving to the
service, we will schedule these jobs for every repo.

Make this step wait a full day between actions if there is no cache
server. This prevents overloading the central server.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>