Scalar.Service: unregister repo when it's not on disk
Update the RepoRegistry to check if the enlistment root
exists on disk before attempting to run the maintenance
verb against it.  If the enlistment root does not exist
on disk, but it's root directory *does* exist, then
also remove the repo from the registry.

Checking that the root of the enlistment path exists
before removing the repo from the registry ensures that
we will not remove repos that are temporarily unavailable
(e.g. due to Bitlocker or a removable drive not being
present). See the discussion in
https://github.com/microsoft/VFSForGit/pull/359 for details.