RepoRegistry: refactor format and access directly from verbs
Previously, the repo registry was owned completely by the
service, and verbs that needed to register repos (e.g.
'scalar clone') would make a request to the service to
perform the registration.

As a consequence of this design, if the service were not
running when repo was cloned, the repo would never be added
to the registry, and the service would never run maintenance
tasks for it.

With the changes in this commit Scalar uses a new approach
for the repo registry:

- The registry is a directory populated with .repo files
- 'scalar clone' will create a new .repo file for its repo (the
  name of the file being the SHA1 hash of its path)
- The .repo file stores the repo's path and owner in JSON format
- Scalar.Service will discover all the registered repos
  by scanning the registry directory each time a maintenance
  task runs

As part of this change, the following code could be removed:

- All registration related messages to/from the service and
  the scalar verbs
- The concept of "active" registered repos.  Now that there's
  not a mount process there's no need to track which repos
  are mounted (i.e. "active")
- All methods in the registry interface that are no longer
  required

Future work:

- Add a new verb for registering and unregistering repos
- Move the '--list-registered' option from the 'service'
  verb to the new verb mentioned above