StatusTests: close file handles eventually
Some tests use OpenFileAndWriteWithoutClose() as a way to check if
our Git commands behave the same. However, if the filestream is
disposed before the command runs, then we can get a situation where the
behavior changes.

By returning a "signalDone" action, we can keep a pointer to the
stream alive, then close it at the appropriate time.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>