ScalarEnlistment: properly search for root with "src"
The previous algorithm that looked for an enlistment without a
.scalar directory did not make a distinction between "src/.git" and
just ".git" as to where the working directory was. Since the
ScalarEnlistment class automatically added a "src" dir to the
enlistment root, that caused maintenance verbs to fail when run
inside the "src" folder (it would add a second "src" dir).

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>