WatchVerb: allow watching vanilla Git repos
This also allows re-registering Scalar repos if they somehow become
unregistered. (We will follow-up with a way to un-watch later.)

The only required update to the maintenance steps right now is to
make the fetch-commits-and-trees step be a no-op for non-Scalar
repos. Eventually, we will want that to run a "git fetch" (in a
way that does not update the refs/remotes space).

Signed-off-by: Derrick Stolee <stolee@gmail.com>