CommitGraphStep: do not immediately delete expired commit-graph files
We have noticed a lot of "cannot find all commit-graph files" errors,
more than is reasonable for simply hard-drive failures. Likely there is
an issue with how Git is comparing the paths of the commit-graph files
it is queueing for deletion and the ones it plans to keep, related to
how we use the --object-dir parameter.

The commit-graph code in Git will "touch" all the files that are
currently in the commit-graph-chain, so as long as we pass a non-zero
expire time, that will prevent us from deleting the files that are
still in the chain despite the path difference.

This is a faster fix than fixing Git first.

I noticed this after running the commit-graph step on a vanilla
Git repo that already had a long commit-graph chain generated by
fetch.writeCommitGraph.

Signed-off-by: Derrick Stolee <stolee@gmail.com>