WindowsScalarService: find a logged-in user on service startup
A big issue with our background maintenance is that the Windows service
currently waits for a login event to find a session id. Without a user
from the session id, we cannot run background maintenance or risk having
admin privileges on all of our repos.

Make an attempt to find a user by enumerating the current active
processes and isolating the active (non-zero) session ids. This allows
background maintenance to start immediately after install or upgrade.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>