FetchStep: clear hidden refs before fetching
If we do nothing, the refs/hidden/* refpsace will grow arbitrarily. This
could cause a problem for users whose remote creates many different ref
names. Such bad behavior requires a user to interact with their refs to
clear (say, "git fetch --prune") but we cannot do the same with the
hidden refs. Since we only need them for the commit-graph writes, delete
the refs before fetching. This will recreate the refs that are on the
remote at that time.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>