Pipelines: use configuration setting for msbuild task
When we use the msbuild task to build the installers from our existing
binaries, we want to make sure we are using the right outputs. Let's
double-check that we are using the release outputs.

It is possible that we are re-building a debug build of scalar.exe in
the "Collect payload files" step, and without a specified version.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>