Publish 'Installers' nupkgs as an Artifact
Publish the 'Scalar.Installers.<platform>' NuGet packages as a build
artifact as well as pushing to the internal feed. This way we can always
derive which package version a particular build pushed.

In VFS for Git we used to keep the build number of the build pipeline
uniform across branches meaning we could always derive the pushed
'Installers' NuGet package version from that build number. Since we now
only include the _product_ version number in branches built matching
`releases/xx.xx.xx` we cannot do that.