Create the correct UpgradeOrchestrator for each platform
During the port to .NET Core 3 and the build system update, an #if
directive was left behind which was being used to select the correct
UpgradeOrchestrator for each OS platform. The result was that upgrade
would always fail on all platforms with a 'NotImplementedException'.

Replace the #if directive with a runtime `IsOSPlatform` check.