Shell out to nuget.exe to verify packages
The NuGet SDK for .NET Core does not support verification of NuGet
packages, even on Windows.

We instead bundle a copy of nuget.exe on Windows installs of Scalar and
shell out to the CLI tool to perform validation.

The tool must be Authenticode signed before we delegate package
verification to it.