Git: remove exact version match
The MinimumGitVersion build variable previously was generated by
opening the Git installer package. Instead, make it a constant
string in the dependencies file. We will update this only when
we make a breaking change.

Modify our version check to always be a minimum version check.

We no longer need the GetGitInstallerVersion build task.

Signed-off-by: Derrick Stolee <stolee@gmail.com>