Windows: wait for processes started as user
When running background maintenance from the service, we want to wait
for the process to finish to guarantee we do not have multiple
maintenance activities running at the same time. However, we have a
custom runner to launch the `scalar run` processes. This runner does
not properly wait for the process to end, so we quickly launch all of
the maintenance activities at once. This can be particularly bad after
opening a laptop after a long sleep.

By adding a wait, we will actually ensure that only one maintenance
activity runs at a time.

Resolves the root cause of #328

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>