ConfigStep: install fsmonitor hook from installed templates
Before, we would copy samples from the .git/hooks/ directory, but that
relies on the timing of when a user initialized their repository.
Further, the .git directory is less safe than the place where Git is
installed. This new mechanism allows us to upgrade the fsmonitor hook
automatically with new Git updates.

Also update core.fsmonitorHookVersion to 2 explicitly.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>