Mac: fully root path to '/usr/local' for hook location
In #349, we updated where the hooks were being copied from. Instead of
the local .git/hooks directory, we copy from the installed templates
directory. This allowws us to grab the latest hook associated with our
Git version.

However, I neglected testing thoroughly on Mac, and the path was not
rooted. This causes the config step to log an error because the path
starts at the current working directory, not from root.

Adjust the ConfigStep to return a 'false' result when this step fails,
which will cause the RunVerbTests to fail if we cannot set the hook as
expected.

The test does not verify the hook is placed because we only place it if
Watchman is installed.

Also, I noticed that the templates directory was wrong! Our installer
puts the templates in a different place than another version of Git.
Perhaps that version was installed by XCode or something, but that
caused the confusion.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>