CommitGraphStep: fix use of --expire-time
See microsoft/git#255. For some reason I got myself confused as to
what --expire-time was for. The command-line interface says one thing
while the internal implementation does something different. Scalar
and VFS for Git were doing what the internal implementation was
expecting, but the option parsing was not properly reflecting the
data correctly.

This means that a lot of users have an excess of commit-graph files
in their object directories. This will quickly clean them all up.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>