StatusStep: run status (and update-index) in the background
By running this occasionally, we will get users out of bad states
where their untracked cache is not updated or their status is a
bit slow. Unlike the status cache in VFS for Git, this does not
trigger in response to file modifications. Instead, it runs every
hour to prevent a user being in a bad state for too long.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>