ConfigStep: disable status.submoduleSummary when using GVFS protocol
A VFS for Git user noticed extremely slow "git status" times, and the
root cause was that they had set status.submoduleSummary=true in their
global config. This causes two "git submodule" subprocesses, which are
very slow. The GVFS protocol does not allow submodules, so disable the
setting when using that protocol.

I investigating fixing this in Git proper, but there is no efficient
way to detect that we can ignore this setting since there could be
an untracked submodule.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>