git-hooks: resolve git symlink on Mac to find hooks
Replace the hardcoded /usr/local/git/ base installation path on macOS
with a `which git` and symlink resolution, to ensure we always find the
correct hooks template directory.

This works for the Homebrew installed Git, or the .pkg installed Git.
Apple-Git would work, but it does not ship with any hook templates
anyway!