CheckGitVersion: improve error messages
A user reported issues with parsing our Git version when cloning. Our
best guess of what happened is that the build machine already had a
version of Git, but it was in a different location whan where our `brew`
formula placed our version of Git. This led the parser to not see a
`.vfs.` in the version string and fail.

Let's make this error more informative. While we are here, split the
failure condition into two distinct cases:

1. We can't find Git anywhere.
2. We can't parse the Git version.

Add the GitBinPath to the other error messages in this method.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>