Watchman: get latest Windows release from GitHub
The old link broke and we never noticed.

The new process downloads a 'watchman.zip' file which expands to
something like 'watchman\watchman-<tag>-windows\' and the *-expansion in
CMD is strange. This requires some careful renaming of directories to
place the results in a consistent place in order for the install script
to work.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>