setup GitHub Actions build and test steps
We define GitHub Actions YAML which runs a simplified version of
the BuildScalarFor*.{bat,sh} and RunUnitTests.{bat,sh} scripts
under Scripts/.

One difference is that we use a Release build configuration
instead of using a Build configuration as our default.

We also do not define the ScalarVersion MSBuild property and
instead rely on what is set in Directory.Build.props.

h/t kyle-rader