remove leftover NeedsServiceVerb test category
In PR #216 extensive work was done to remove some of the repository
registry support of the Scalar service, and many tests no longer
required the MacTODO.NeedsServiceVerb category flag and could be
successfully executed on the Mac.

One test which appears to have been missed was the
SecondCloneSucceedsWithMissingTrees() test in SharedCacheTests.cs;
however, it also succeeds on the Mac now and so we can remove
this functional test category entirely.