remove leftover NeedsScalarConfig test category
In the VFSForGit project, support for upgrading was added in
stages, and at first the Mac platform did not have a full
implementation of the local config class and "config" verb,
so the relevant tests were marked to be excluded on the Mac.

Subsequently, full support for the "config" verb was added,
but the tests were not enabled on the Mac.  So this test
category is no longer meaningful, and the tests would succeed
on the Mac if several other changes were made (i.e., the
NeedsUpdatesForNonVirtualizedMode category was also removed
from these tests, and the functional test suite is run as
root).

While there may not be an intention to ever run these tests,
at a minimum we can simplify the test configuration by
removing this category.

See microsoft/VFSForGit@77fa9962a2b0da2255aebae4e020f4c95dc79219
and microsoft/VFSForGit@75115280eb9338103317e78eee4a3aa3b0184e9f.