remove Scalar_Clone.sh utility script on Mac
This utility script was originally added to the VFSForGit project
in the early bulk commit
microsoft/VFSForGit@785ccb4a67281d0eb10f19cb53e6431b62b88555, as part
of a trio of scripts to make it easier to mount, clone, and unmount
a test repository in the ~/GVFSTest directory.

As part of Scalar PR #216, though, the mount and unmount helper
scripts were removed as no longer relevant.

While the remaining script is still functional, its utility is
limited as its restriction to a fixed ~/ScalarTest directory
makes less sense without the matching mount/unmount scripts, so
we remove it here.