Actions: run functional tests on all platforms
To run the functional test suite in GitHub Actions, we
write platform-specific steps to install the GVFS Git package,
but then try to use a single common step to run the functional
tests themselves.

The commands in the Git installer steps are drawn in part
from those found in the existing
Scalar.Installer.{Mac,Windows}/InstallScalar.template.{sh,bat}
scripts.

On Windows, we specifically need to ensure that our version of
Git is the one in "C:\Program Files\Git\cmd\git.exe", so we start
install process by uninstalling the GitForWindows version provided
on the Actions instance.  For future debugging purposes we also
include a final Get-ItemProperty PowerShell query to report
whether we successfully updated the GitForWindows registry entry.

We also disable the fast-fail option on our matrix builds so
all platforms will run to completion, allowing us to see
any mix of platform-specific errors without short-circuting
due to the first CI failure on any platform.

Note that we do not yet install Watchman or provide additional
functional test runs using it; that work remains to be done.