POSIXPlatforms: use base paths from env vars
To provide partial support for the Linux XDG Base Directory
Specification, and in particular to locate our cache, config,
and data directories for Scalar under Linux-standard locations
such as ~/.cache/scalar, we introduce a common
TryGetEnvironmentVariableBasePath() method in the shared
POSIXPlatform, which we then use as appropriate for each
type of user-specific directory.

This method accepts a list of pairs of environment variables
and path components, which are checked in sequence until a
non-empty environment variable is found; its value is then
returned with the corresponding path components appended.
Thus we can try to return "$XDG_DATA_HOME/scalar", but if that
variable is undefined or empty, return "$HOME/.local/scalar"
instead.

We also update the relevant Mac and Linux methods and accessors
to utilize TryGetEnvironmentVariableBasePath() with the
appropriate pairs of environment variables and paths.  On macOS,
we update the TryGetDefaultLocalCacheRoot() and
GetDataRootForScalarImplementation() methods; on Linux
we also update ScalarConfigPath().  We make no functional change
to the paths returned on macOS; however, errors due to an
undefined $HOME variable or other exceptions are now always
reported.  On Linux, we implement the user-specific paths from
the XDG specification.

Note that on Linux we do not yet create missing directories with
a 0700 mode, as specified by the XDG standard, and we also do not
check system-wide locations from either the $XDG_{CONFIG,DATA}_DIRS
environment variables or the defaults for them.  The latter
might require more extensive refactoring to distinguish cases
in which user-specific data or configurations are to be written,
as these should possibly not be written to system-wide files.