FunctionalTests: rename syscall wrappers for POSIX
We adjust the name of the MacCreateHardLink() wrapper for the
link(2) syscall to be POSIXCreateHardLink(), as this also runs
on Linux.

We also create distinct MacChmod() and LinuxChmod() wrappers
for the chmod(2) syscall as the size of the mode argument
varies between platforms.

From microsoft/VFSForGit@79ef5723656a110f9b1e3cd76a273d710420ec95.