FunctionalTests: remove ExtraCoverage category
A limited subset of functional tests are marked as belonging to
the ExtraCoverage category, which was introduced in the
microsoft/VFSForGit#1046 PR as an enhancement over the
FullSuiteOnly category from the Azure Repos PR 302803.
The intent was to make the default functional test run faster
by excluding some less-fragile tests.

However, in Scalar, many of these tests are also marked with
NeedsUpdatesForNonVirtualizedMode and therefore do not run at all,
while the ones which remain have been running in CI on the macOS
platform but not Windows, because Scripts/Mac/RunFunctionalTests.sh
passes the --full-suite option.

In order to simplify the CI and functional test framework,
we eliminate the ExtraCoverage category entirely, which
ensures all working tests run on all platforms in CI.