FunctionalTests: remove --windows-only option
Since the number of Windows-only functional tests is quite small,
and we generally want to run all tests in CI, the original purpose
of the --windows-only option to the functional test suite has
less value than in the VFSForGit project.  We also do not have
an equivalent --mac-only or --linux-only option.

Therefore we simply remove this option, which tidies the program
logic and ensures all platforms have an equal set of test options.