Scripts/Mac: drop --full-suite functional test opt
In order to align better with the Windows RunFunctionalTests.bat
script, on macOS we change the RunFunctionalTests.sh script so it
does not always pass the --full-suite option to the functional
test program.

This has no substantive impact right now, as the --full-suite option
just defines a set of FileSystemRunners applicable to the platform
in ScalarTestConfig.FileSystemRunners, which is then returned by
FileSystemRunner.Runners() and may be used to parameterize tests
so they run with each type of runner.

However, at present, the only tests parameterized in this way are
the EnlistmentPerFixture.GitMoveRenameTests, which are also in
the NeedsUpdatesForNonVirtualizedMode category and thus do not
run at all.