FunctionalTests: support variable local cache root
We update the CloneWithDefaultLocalCacheLocation() functional
test so it accommodates variable local cache paths on Linux,
depending on whether $XDG_CACHE_HOME is set or not.  This allows
the test to succeed on developer systems where we want to avoid
polluting the default ~/.cache/scalar directory during test runs
and so set $XDG_CACHE_HOME explicitly.

Note that the new ScalarTestConfig.DefaultLocalCacheRoot accessor
returns a path on Windows which should be "C:\.scalarCache",
to match what would be supplied by the TryGetDefaultLocalCacheRoot()
method of the WindowsPlatform if it were passed the default
functional test enlistment root path.  However, this Windows-
specific DefaultLocalCacheRoot value is unused, as the only
functional test which uses it is CloneWithDefaultLocalCacheLocation()
and that test belongs to the POSIXOnly category.