Platforms: validate case-sensitivity in clone dir
All OS platforms are capable of supporting directories with
case-sensitivity of an opposite mode to their conventional
one (i.e., case-sensitive on Windows/Mac or case-insensitive
on Linux).  However, Scalar at present assumes the normal
case-sensitivity of each platform.

Therefore we expand the check currently implemented only in
the MacFileSystem IsFileSystemSupported() method so it
executes on all platforms and tests whether the directory
into which we are cloning has the expected case-sensitivity.

(The IsFileSystemSupported() method is only called by the
CloneVerb command, so we only execute this check at the time
of initial cloning.)