FunctionalTests: refactor and refine CloneTests
Prior to adding any tests which check path case-sensitivity,
we refactor the existing CloneTests and make several minor
refinements.

The SubfolderCloneShouldFail() private method is no longer
utilized by more than one test, as it was previously until
commit e0c16fb264bf11a59eb2a35a8b785fcb9979ac70 which removed
the CloneInsideUnmountedEnlistment() test.  Therefore we
remove this utility method and instead create a new
RunCloneCommand() private method which is more generic and
which we can use in several of the remaining tests to eliminate
common ProcessHelper setup code.

We standardize in RunCloneCommand() on using the repository
URL from ScalarTestConfig.RepoToClone, which is normally
the Default.RepoToClone value from the
Scalar.FunctionalTests.Properties.Settings class, but which
can be overridden using the --repo-to-clone command-line
option to the functional test program.  One test was already
using this value, but two others used the default one only.

In the CloneInsideExistingEnlistment() test we also replace a
Windows-specific enlistment root path which used backslash path
separators with one created using Path.Combine().