FunctionalTests: add multi-OS fetch cleanup tests
We add three tests to the FetchStepWithoutSharedCacheTests
suite, one of which is based on the existing
FetchStepCleansUpBadPrefetchPack() test but which checks
the behaviour when the bad prefetch pack files' names
differ in case from the expected ones.

On Linux, where we expect case-sensitive filesystems, we
ignore such bad files, while on macOS and Windows we treat
them identically as in the existing test and delete them.

The other two new tests perform similar checks but for the
*.keep sentinel files that are created to mark the newest good
prefetch packfile.  Since a Scalar fetch maintenance job
is expected to clean these up, leaving only one such file,
we test that is true in FetchStepCleansUpBadKeepFiles(), and
then our other new test performs the same case-sensitive variation
as described above for FetchStepCleansUpBadPrefetchPack().