FetchStep: ignore .keep files with wrong prefix
When listing all of the *.keep files in order to delete all but
the one corresponding to the most recent good prefetch packfile,
we ignore any whose names do not start with our "prefetch" prefix.

This aligns with the logic expressed earlier in the same
UpdateKeepPacks() method where *.pack files are searched, but
only if their name starts with "prefetch-" and the expected
maximum timestamp.