ConfigStep: launch background maintenance through Git
On Linux, when using Git v2.28.0.vfs.1.0 or later (and this will be
adjusted with the v2.29.0.vfs.0.0 release), we have access to the
cron-enabled background maintenance in Git. This enables background
maintenance for Scalar on Linux.

While we _could_ use this integration on macOS, we will still ship with
Scalar.Service on that platform for now to minimize possible issues.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>