Platforms: undo partial Linux/MacPlatform classes
The {Linux,Mac}Platform.cs and {Linux,Mac}Platform.Shared.cs
classes exist as pairs of partial classes in order to share
static methods in the Shared classes with the GVFS.Hooks project;
however, that project was removed in commit
f12effc371d7b1457ab58c6d44959632404e5998 in PR #23, so we can
merge the implementation of several methods back into the common
POSIXPlatform class and remove the shared ones.

We also align all the per-platform EnvironmentVariables arrays
to be private.