Platforms: undo partial Windows classes
The WindowsPlatform.cs and WindowsPlatform.Shared.cs classes exist
as a pair of partial classes in order to share the static methods
in the latter with the GVFS.Hooks project; however, that project
was removed in commit f12effc371d7b1457ab58c6d44959632404e5998
in PR #23, so we can merge the implementation of several methods
back into the main WindowsPlatform class and remove the shared one.

We also slightly reorder the GetUpgradeProtectedDataDirectory()
method so it aligns with those in ScalarPlatform and POSIXPlatform,
which may help with future refactoring or simplification.