Platforms: remove unused active process check
The last caller of the IsProcessActive() methods in the
ScalarPlatform classes was removed in commit
cedeeaa3ba68e86d4aa0ec3e5d453b8c033f7f6b in PR #29.

Therefore we can remove this method as well as the native
Windows system call wrappers and enums which were used only
by the WindowPlatform implementation of this method.