clone: respect remote default branch name
Attempt to determine the remote's default branch name when performing a
`scalar clone` using the vanilla Git protocols (not using the GVFS helper).

Use ls-remote to lookup the HEAD ref on the remote, and parse the
results. If the remote's HEAD is not a branch, or we fail to parse the
output for any reason, consult Git itself for the default branch name as
a fallback.