RunVerb: pass git features to ConfigStep
The ConfigStep makes a decision about calling 'git maintenance start'
depending on the platform (Linux only for now) and the Git version.
However, the code for computing the feature flags is in ScalarVerb and
needs to be injected by passing it through the ConfigStep constructor.
This was missed in the earlier work that inserted this into other
maintenance steps.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>