ConfigStep.cs: use feature.scalar to enable Watchman or the Builtin FSMonitor
Extend definition of `feature.scalar` to select either Watchman or the new
builtin FSmonitor.

Update the `ConfigStep.cs:TrySetConfig()` to set/unset `core.useBuiltinFSMonitor`
when enabling/disabling the builtin `git fsmonitor--daemon`.  This feature is
new as of Git v2.31.0.vfs.0.0.

Previous beta versions of Git used `core.fsmonitor == :internal:` to select it.

Also, coordinate values of `core.fsmonitor` with `core.useBuiltinFSMonitor`
so that we can select neither or one or the other.

Signed-off-by: Jeff Hostetler <jeffhost@microsoft.com>