ConfigStep: Update config for core.useBuiltinFSMonitor
Update the `ConfigStep.cs:TrySetConfig()` to set/unset `core.useBuiltinFSMonitor`
when enabling/disabling the builtin `git fsmonitor--daemon`.  This is the new
value as of v2.31.0.

Previous beta versions of Git used `core.fsmonitor == :internal:` to select it.

Also, coordinate values of `core.fsmonitor` with `core.useBuiltinFSMonitor`
so that we can select neither or one or the other.

Signed-off-by: Jeff Hostetler <jeffhost@microsoft.com>