ConfigStep: commitGraph.generationVersion=1
The generation number v2 is better for algorithmic reasons, but it is
backwards-compatible with version 1 and in doing so requires reading
more data from disk. That extra data load causes a performance
degradation in the typical case.

Let's avoid that for now, until a non-backwards-compatible version is
ready that avoids this performance problem.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>